Guarantee delivers pokemon on 200 response
